refactor: share current error diagnostics
The feedback UI needs to render the notebook's current errors, but the
only extractor lived inside the AI context provider and was entangled
with AI-specific formatting that embeds cell source code. Sharing that
provider wholesale would leak source into issue reports.

This moves the neutral extraction (error parsing, traceback resolution,
`getCellErrorEntries`, and the per-error describe/format helpers) into
`core/errors/error-entries` and adds `formatCellError`, an issue-safe
formatter that emits the cell name and plain-text traceback but never
the cell code. The AI provider keeps its own `formatCellErrorDetails`,
so AI context still includes `Code:` and datasource schema; its output
is unchanged.
